Hackish check for the case where FB sends "count: 0" and no "data:"
Mogli, when creating new Mogli::* objects, can be passed:

-- a hash of key => value pairs, which it will pass to new() to create an object;
-- an array of hashes, from which it will create a list of objects;
-- a hash containing {:count => X, :data => [...]} which will climb down into
the "data" value to get the list of hashes to make a list of objects from.

In the case where there are no comments on a post (possibly in any case where
:count is zero), Facebook does not send a :data key -at-all-.  Mogli then
faithfully detects the first case, and tries to make a single comment out
of {:count => 0}.

This detects that case and steps around it.